deleted

+++ app/Views/tasks/tasksForm.php
@@ -9,8 +9,10 @@
                 <h4 class="text-center">Task Creation</h4>
             </div>
             <div class="col-sm-3">                 
-                <button type="button" class="btn btn-primary float-end " onclick="history.back();" ><i class="fa fa-arrow-left"></i></button> 
-                <a href="tasks/delete/<?= $task->task_id ?>" class="btn btn-danger float-end me-3" onclick="return confirm('Are you sure you want to delete this record?');"><i class="fa fa-trash"></i></a>
+                <button type="button" class="btn btn-primary float-end " onclick="history.back();" ><i class="fa fa-arrow-left"></i></button>
+                <?php if($mode != 'create') : ?> 
+                    <a href="tasks/delete/<?= $task->task_id ?>" class="btn btn-danger float-end me-3" onclick="return confirm('Are you sure you want to delete this record?');"><i class="fa fa-trash"></i></a>
+                <?php endif; ?>
                 <button type="button" class="btn btn-primary float-end me-3"><i class="fa fa-edit"></i></button>
             </div>
         </div>
